aisd_lab5_rbt/src/main/java/pl/edu/pw/ee/RbtMap.java
package pl.edu.pw.ee;

import pl.edu.pw.ee.services.MapInterface;

public class RbtMap<K extends Comparable<K>, V> implements MapInterface<K, V> {

    private final RedBlackTree<K, V> tree;

    public <K, V> RbtMap() {
        tree = new RedBlackTree<>();
    }

    public int getNElems() {
        return tree.getNElems();
    }

    public void deleteMax() {
        tree.deleteMax();
    }

    @Override
    public void setValue(K key, V value) {
        if (key == null || value == null) {
            throw new IllegalArgumentException("Params (key, value) cannot be null.");
        }
        tree.put(key, value);
    }

    @Override
    public V getValue(K key) {
        if (key == null) {
            throw new IllegalArgumentException("Cannot get value by null key.");
        }
        return tree.get(key);
    }

}
